+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 4.2.0
+## 4.2.0 (November 25, 2023)
 * Exit process when there are no input files ([#83](https://github.com/vseventer/sharp-cli/issues/83)).
 * Added `unflatten` command.
 * Added `extendWith` option to `extend` command.